src/main/java/org/ruhlendavis/mumue/connection/stages/loginphase/WaitForLoginIdStage.java
package org.ruhlendavis.mumue.connection.stages.loginphase;

import org.ruhlendavis.mumue.configuration.Configuration;
import org.ruhlendavis.mumue.connection.Connection;
import org.ruhlendavis.mumue.connection.stages.ConnectionStage;

public class WaitForLoginIdStage implements ConnectionStage {
    @Override
    public ConnectionStage execute(Connection connection, Configuration configuration) {
        if (connection.getInputQueue().isEmpty()) {
            return this;
        }
        return new PasswordPromptStage();
    }
}
